VBO-indeksi on nyt muuttujatyyppiä `GLuint`, ei `unsigned short`.
Indeksejä voi olla 2^32 = 4294967296 aiemman 65536:n sijaan.

	modified:   common/vboindexer.cpp
	modified:   common/vboindexer.hpp
	modified:   several_objects/several_objects.cpp
	modified:   ylikuutio/ylikuutio.cpp

+++ common/vboindexer.hpp
@@ -1,3 +1,6 @@
+// Include GLEW
+#include <GL/glew.h>
+
 #ifndef VBOINDEXER_HPP
 #define VBOINDEXER_HPP
 
@@ -6,7 +9,7 @@ void indexVBO(
 	std::vector<glm::vec2> & in_uvs,
 	std::vector<glm::vec3> & in_normals,
 
-	std::vector<unsigned short> & out_indices,
+	std::vector<GLuint> & out_indices,
 	std::vector<glm::vec3> & out_vertices,
 	std::vector<glm::vec2> & out_uvs,
 	std::vector<glm::vec3> & out_normals
@@ -20,7 +23,7 @@ void indexVBO_TBN(
 	std::vector<glm::vec3> & in_tangents,
 	std::vector<glm::vec3> & in_bitangents,
 
-	std::vector<unsigned short> & out_indices,
+	std::vector<GLuint> & out_indices,
 	std::vector<glm::vec3> & out_vertices,
 	std::vector<glm::vec2> & out_uvs,
 	std::vector<glm::vec3> & out_normals,
